Move Cleanup.h from include/tscpp/api to include/ts.
Change the namespace of its declarations from atscppapi to tsapi::c_support.
This is in support of removing the main, high-overhead parts of the ATS C++ API.
