+++ tests/gold_tests/pluginTest/tsapi/test_tsapi.cc
@@ -26,9 +26,9 @@
 #include <bitset>
 
 #include <tscpp/util/PostScript.h>
-#include <tscpp/api/Cleanup.h>
 
 #include <ts/ts.h>
+#include <ts/Cleanup.h>
 #include <ts/remap.h>
 
 /*
@@ -41,10 +41,10 @@ namespace
 {
 char PIName[] = PINAME;
 
-atscppapi::TSDbgCtlUniqPtr dbg_ctl_guard{TSDbgCtlCreate(PIName)};
+tsapi::c_support::TSDbgCtlUniqPtr dbg_ctl_guard{TSDbgCtlCreate(PIName)};
 TSDbgCtl const *const dbg_ctl{dbg_ctl_guard.get()};
 
-atscppapi::TSDbgCtlUniqPtr off_dbg_ctl_guard{TSDbgCtlCreate("yada-yada-yada")};
+tsapi::c_support::TSDbgCtlUniqPtr off_dbg_ctl_guard{TSDbgCtlCreate("yada-yada-yada")};
 TSDbgCtl const *const off_dbg_ctl{off_dbg_ctl_guard.get()};
 
 // NOTE:  It's important to flush this after writing so that a gold test using this plugin can examine the log before TS
